moves computation of merkle "zero hashes" to build time
- in order to drop use of `lazy_static` dep which is not easy to use
  in `no-std` context
