+++ ssz-rs/Cargo.toml
@@ -24,7 +24,6 @@ thiserror = "1.0.25"
 bitvec = "1.0.0"
 ssz-rs-derive = { path = "../ssz-rs-derive"}
 sha2 = "0.9.8"
-lazy_static = "1.4.0"
 serde = { version = "1.0", features = ["derive"], optional = true }
 hex = {version = "0.4.3", optional = true }
 num-bigint = "0.4.3"
@@ -34,3 +33,6 @@ hex-literal = "0.3.3"
 snap = "1.0"
 project-root = "0.2.2"
 serde_json = "1.0.81"
+
+[build-dependencies]
+sha2 = "0.9.8"
